+++ CloudVisualization/CloudVisualization.csproj
@@ -9,4 +9,8 @@
       <PackageReference Include="System.Drawing.Common" Version="5.0.0" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="Core" />
+    </ItemGroup>
+
 </Project>
