+++ RectanglesCloudLayouter/Interfaces/ICloudLayouter.cs
@@ -1,4 +1,4 @@
-﻿namespace CloudLayouter.Interfaces
+﻿namespace RectanglesCloudLayouter.Interfaces
 {
     public interface ICloudLayouter
     {
