+++ RectanglesCloudLayouter/Interfaces/ISpiral.cs
@@ -1,6 +1,6 @@
 ﻿using System.Drawing;
 
-namespace CloudLayouter.Interfaces
+namespace RectanglesCloudLayouter.Interfaces
 {
     public interface ISpiral
     {
