+++ RectanglesCloudLayouter/RectanglesCloudLayouter.csproj
@@ -4,4 +4,8 @@
         <TargetFramework>netcoreapp3.1</TargetFramework>
     </PropertyGroup>
 
+    <ItemGroup>
+      <PackageReference Include="System.Drawing.Common" Version="5.0.0" />
+    </ItemGroup>
+
 </Project>
