+++ RectanglesCloudLayouterTests/PointsDistance_Should.cs
@@ -1,10 +1,9 @@
 ﻿using System.Drawing;
-using CloudLayouter;
-using CloudLayouter.SpecialMethods;
 using FluentAssertions;
 using NUnit.Framework;
+using RectanglesCloudLayouter.SpecialMethods;
 
-namespace CloudLayouterTests
+namespace RectanglesCloudLayouterTests
 {
     [TestFixture]
     public class PointsDistanceShould
