+++ TagsCloudContainer/WordsFrequency/WordsFrequency.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 
 namespace TagsCloudContainer.WordsFrequency
 {
