+++ di.sln
@@ -2,11 +2,9 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FractalPainter", "FractalPainter\FractalPainter.csproj", "{4D70883B-6F8B-4166-802F-8EDC9BE93199}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CloudLayouter", "CloudLayouter\CloudLayouter.csproj", "{282C70CC-1DD1-4B95-92AF-7D42E9418388}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RectanglesCloudLayouter", "RectanglesCloudLayouter\RectanglesCloudLayouter.csproj", "{282C70CC-1DD1-4B95-92AF-7D42E9418388}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CloudVisualization", "CloudVisualization\CloudVisualization.csproj", "{568077D1-049C-4A99-903D-50C7248D3A7F}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CloudLayouterTests", "CloudLayouterTests\CloudLayouterTests.csproj", "{A8F663CF-3D85-415A-BB4C-97579B5D66B8}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RectanglesCloudLayouterTests", "RectanglesCloudLayouterTests\RectanglesCloudLayouterTests.csproj", "{A8F663CF-3D85-415A-BB4C-97579B5D66B8}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagsCloudContainerTests", "TagsCloudContainerTests\TagsCloudContainerTests.csproj", "{AE9A8D9C-4AE6-4918-8941-82A155CA0B04}"
 EndProject
@@ -26,10 +24,6 @@ Global
 		{282C70CC-1DD1-4B95-92AF-7D42E9418388}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{282C70CC-1DD1-4B95-92AF-7D42E9418388}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{282C70CC-1DD1-4B95-92AF-7D42E9418388}.Release|Any CPU.Build.0 = Release|Any CPU
-		{568077D1-049C-4A99-903D-50C7248D3A7F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{568077D1-049C-4A99-903D-50C7248D3A7F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{568077D1-049C-4A99-903D-50C7248D3A7F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{568077D1-049C-4A99-903D-50C7248D3A7F}.Release|Any CPU.Build.0 = Release|Any CPU
 		{A8F663CF-3D85-415A-BB4C-97579B5D66B8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A8F663CF-3D85-415A-BB4C-97579B5D66B8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A8F663CF-3D85-415A-BB4C-97579B5D66B8}.Release|Any CPU.ActiveCfg = Release|Any CPU
